style:moved hooks and constants to utils
